+++ CHANGELOG.md
@@ -10,7 +10,7 @@ CHANGELOG
 - **Changed** `IJ2TPL.escapeHTML` -> `IJ2TPL.escape`
 - Fixed check on **empty array**(difference with mustache.js)
 - Supported partial template
-- Optimized tokenize for comments
+- Optimized `IJ2TPL.tokenize` for comments
 - ...
 
 status: **DEVELOPING**